+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParameters.kt
@@ -1,0 +1,6 @@
+package at.asitplus.openid
+
+interface RequestParameters
+
+
+