+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/Json.kt
@@ -1,5 +1,6 @@
 package at.asitplus.wallet.lib.oidvci
 
+import at.asitplus.wallet.lib.data.vckJsonSerializer
 import kotlinx.serialization.json.Json
 
 val jsonSerializer by lazy {
@@ -8,5 +9,6 @@ val jsonSerializer by lazy {
         encodeDefaults = false
         ignoreUnknownKeys = true
         isLenient = true
+        serializersModule = vckJsonSerializer.serializersModule
     }
 }